Conflicts:
	modules/juce_audio_plugin_client/AU/juce_AU_Wrapper.mm
	modules/juce_audio_processors/processors/juce_AudioProcessor.cpp
	modules/juce_audio_processors/processors/juce_AudioProcessor.h

+++ modules/juce_audio_formats/format/juce_AudioFormatReader.cpp
@@ -199,7 +199,7 @@ void AudioFormatReader::readMaxLevels (int64 startSampleInFile, int64 numSamples
     }
 
     const int bufferSize = (int) jmin (numSamples, (int64) 4096);
-    AudioSampleBuffer tempSampleBuffer (numChannels, bufferSize);
+    AudioSampleBuffer tempSampleBuffer ((int) numChannels, bufferSize);
 
     float** const floatBuffer = tempSampleBuffer.getArrayOfChannels();
     int* const* intBuffer = reinterpret_cast<int* const*> (floatBuffer);
@@ -219,7 +219,7 @@ void AudioFormatReader::readMaxLevels (int64 startSampleInFile, int64 numSamples
 
             numSamples -= numToDo;
             startSampleInFile += numToDo;
-            getStereoMinAndMax (floatBuffer, numChannels, numToDo, lmin, lmax, rmin, rmax);
+            getStereoMinAndMax (floatBuffer, (int) numChannels, numToDo, lmin, lmax, rmin, rmax);
         }
 
         lowestLeft   = lmin;
@@ -242,7 +242,7 @@ void AudioFormatReader::readMaxLevels (int64 startSampleInFile, int64 numSamples
 
             numSamples -= numToDo;
             startSampleInFile += numToDo;
-            getStereoMinAndMax (intBuffer, numChannels, numToDo, lmin, lmax, rmin, rmax);
+            getStereoMinAndMax (intBuffer, (int) numChannels, numToDo, lmin, lmax, rmin, rmax);
         }
 
         lowestLeft   = lmin / (float) std::numeric_limits<int>::max();
